# docs/API_SPECIFICATION.yaml
openapi: 3.0.0
info:
  title: ClearDeed API
  version: 1.0.0
  description: End-to-end verified real estate & investment execution platform
  contact:
    name: ClearDeed Advisors LLP
servers:
  - url: https://api.cleardeed.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Development
paths:
...
  /referral-partners:
    post:
      summary: Register as referral partner (admin creates)
      tags:
        - Referral Partners
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile_number:
                  type: string
                full_name:
                  type: string
                email:
                  type: string
                city:
                  type: string
                partner_type:
                  type: string
                  enum: ['agent', 'verified_user']
                agency_name:
                  type: string
                agent_license_number:
                  type: string
      responses:
        '201':
          description: Referral partner registered
  /referral-partners/{partner_id}:
    get:
      summary: Get referral partner details
      tags:
        - Referral Partners
      security:
        - BearerAuth: []
      parameters:
        - name: partner_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Partner details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferralPartner'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReferralPartner:
      type: object
      properties:
        id:
          type: integer
        mobile_number:
          type: string
        full_name:
          type: string
        partner_type:
          type: string
        status:
          type: string
        commission_enabled:
          type: boolean
        total_commission_earned:
          type: number
...